requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-mistralai
langchain-huggingface
python-dotenv
pydantic
streamlit